backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6
anthropic==0.30.0
openai==1.30.0
python-dotenv==1.0.0
requests==2.31.0